Fri, Oct 13, 2023, 1:59 AM +05:00

+++ hours/hours.c
@@ -42,7 +42,7 @@ float calc_hours(int hours[], int weeks, char output)
 
     else if (output == 'A')
     {
-        int average = total_hours / weeks;
+        float average = total_hours / weeks;
         return average;
     }
     else
